src/tools/requirements.txt
azure-storage-blob
azure-storage-queue
pillow